+++ plutarch-api/CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
+## 2.1.1 -- 22-01-2024
+
+### Added
+
+* `plookup`, `plookupData`, `plookupDataWith` to `Plutarch.Api.AssocMap`
+
 ## 2.1.0 -- 22-01-2024
 
 ### Added