+++ plutarch-extra/CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
+## 1.3.1 -- 22-01-2024
+
+### Changed
+
+* `Plutarch.Extra.Interval`, `Plutarch.Extra.Map`, and `Plutarch.Extra.Maybe` use 
+   `plutarch-api` definitions
+
+### Removed
+
+* `Plutarch.Extra.Api` as these functions are now in `plutarch-api`
+
 ## 1.2.1 -- 17-01-2024
 
 ### Fixed